% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_assesment.R
\name{calculate_detection_regularity}
\alias{calculate_detection_regularity}
\title{Calculate detection regularity (CV of daily detection counts)}
\usage{
calculate_detection_regularity(detection_data)
}
\description{
Calculate detection regularity (CV of daily detection counts)
}
\keyword{internal}
