% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_skewness}
\alias{calculate_skewness}
\title{Calculate skewness of a numeric vector}
\usage{
calculate_skewness(x)
}
\description{
Calculate skewness of a numeric vector
}
\keyword{internal}
