% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_assesment.R
\name{calculate_species_accumulation_rate}
\alias{calculate_species_accumulation_rate}
\title{Calculate species accumulation rate}
\usage{
calculate_species_accumulation_rate(detection_data)
}
\description{
Calculate species accumulation rate
}
\keyword{internal}
