% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{create_anomaly_plot}
\alias{create_anomaly_plot}
\title{Create anomaly score visualization}
\usage{
create_anomaly_plot(metrics_df)
}
\description{
Create anomaly score visualization
}
\keyword{internal}
