% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{create_confidence_plot}
\alias{create_confidence_plot}
\title{Create confidence interval distribution plot}
\usage{
create_confidence_plot(metrics_df, has_scales = FALSE)
}
\description{
Create confidence interval distribution plot
}
\keyword{internal}
