% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{create_overview_plot}
\alias{create_overview_plot}
\title{Create overview scatter plot of key metrics}
\usage{
create_overview_plot(metrics_df, has_viridis = FALSE)
}
\description{
Create overview scatter plot of key metrics
}
\keyword{internal}
