% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{create_spatial_plot}
\alias{create_spatial_plot}
\title{Create spatial distribution plot}
\usage{
create_spatial_plot(detection_data, metrics_df, species_name = NULL)
}
\description{
Create spatial distribution plot
}
\keyword{internal}
