% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{create_temporal_plot}
\alias{create_temporal_plot}
\title{Create temporal activity pattern plot}
\usage{
create_temporal_plot(detection_data, species_name = NULL)
}
\description{
Create temporal activity pattern plot
}
\keyword{internal}
