% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamPatterns-class.R
\name{plot.pam_metrics}
\alias{plot.pam_metrics}
\title{Plot method for pam_metrics objects}
\usage{
\method{plot}{pam_metrics}(x, type = "overview", ...)
}
\arguments{
\item{x}{A pam_metrics object}

\item{type}{Character. Type of plot to create. See \code{\link{plot_pattern_diagnostics}}}

\item{...}{Additional arguments passed to plot_pattern_diagnostics}
}
\description{
Plot method for pam_metrics objects
}
