% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamPatterns-class.R
\name{print.pam_metrics}
\alias{print.pam_metrics}
\title{Print method for pam_metrics objects}
\usage{
\method{print}{pam_metrics}(x, n = 10, ...)
}
\arguments{
\item{x}{A pam_metrics object}

\item{n}{Number of rows to print (default 10)}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for pam_metrics objects
}
