% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamPatterns-class.R
\name{print.pam_metrics}
\alias{print.pam_metrics}
\title{Print method for pam_metrics objects}
\usage{
\method{print}{pam_metrics}(x, n = 10, ...)
}
\arguments{
\item{x}{A pam_metrics object}

\item{n}{Integer, number of rows to print from the head of the data frame.
Default is 10. Use `NULL` or `Inf` to print all rows.}

\item{...}{Additional arguments passed to the underlying print method for data frames.}
}
\description{
Print method for pam_metrics objects
}
