% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamPatterns-class.R
\name{summary.pam_metrics}
\alias{summary.pam_metrics}
\title{Summary method for pam_metrics objects}
\usage{
\method{summary}{pam_metrics}(object, ...)
}
\arguments{
\item{object}{A pam_metrics object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for pam_metrics objects
}
